Fix server error by switching to gun@0.2.5, use express to simplify server, rename files/folders for more secure serving and structural clarity.
The server was throwing an error because Gun().attach was undefined. We struggled with it for quite some time before realizing it was due to cutting edge gun releases from the dev branch on github. We switched to gun and that solved the problem (change it back as soon as things are stable/tested)

There was a lot of extra unnecessary code to serve client requests. We brought in express to simplify things and reduce the amount of troubleshooting needed to solve the server error (mentioned above). Mission accomplished!

Renamed http.js to server.js and updated the references to reflect that.