Fix failing nightwatch test, code cleaning.
There were a few issues in the third nightwatch test that have been resolved. For one, we were trying to access a local scope node variable on the client through `browser.execute`.  That's been fixed by passing an argument through the second parameter (see API docs). There was another statement that was contradictive (I believe intentionally) that both expected an element to be visible and not to be visible simultaniously.
In chat.js, the username.get() function was prematurely returning false as the value every time, causing the username to always be "false" and failing the tests.
Removed some troubleshooting comments and cleaned up some code. Yay!