Add default launch url, remove chai from tests, rename chat.spec, dry up code.
Added localhost:4242 as the default launch url and removed all references in the nightwatch tests, instead using the `.init()`alias.
Instead of matching against the title, nightwatch now ensures that the server is on by checking for elements on the page. This allows for more accurate reporting in the nightwatch CLI, removal of the chai expect library, and more predictable/consistent looking tests.

Since the username logic was abstracted into it's own file, the `chat.spec.js` name didn't make much sense, as it really just tests the username logic. As such I've renamed the file to reflect this: `username.spec.js`.